finagle-core: Improve handling of unavailable nodes when using deterministic aperture
Problem

We want to ensure that Aperture can deal with down
nodes efficiently across configuration. The default
configuration (when deterministic ordering is off)
shuffles all the healthy nodes to the front of the aperture.
This works really well and we can use the same model when
ordering matters, so long as we return to the original order
when the transient failure is gone.

Note, we also cannot keep the ordering intact and simply expand
the aperture. For example, we can't rely on the load band controller
to quickly widen the aperture and avoid down nodes since,
by definition, it tries to dampen the rate of changes to the
aperture.

Solution

Shuffle the healthy nodes to the front and keep track
of the busy ones. When the busy nodes recover, trigger
a rebuild.

Result

All aperture configurations can sustain a rolling restart
batch size of at least 40% percent of the cluster while
maintaining a success rate of 100%, if the backend capacity
permits.

RB_ID=914653