finagle-*: Refactor HttpNackFilter and expose its Role
Problem

The HttpNackFilter is private[finagle] so users don't have a way of
accessing its Role, forcing them to make their own from the String name
which is a private implementation detail.

Solution

Expose the Role. While doing so, we clean up some of the nack related
constants defined in HttpNackFilter but duplicated in finagle-mux.

RB_ID=914828