finagle-core: Fix connection leak in ListeningStackServer
Problem / Solution

There is a race condition in Listening stack server between a new
connection being accepted and registered with the server and the server
being closed. We fix it by ensuring that all connections that attempt to
register themselves after the server has been closed are immediately
closed with the passed deadline.

RB_ID=914846