finagle-netty4: Add ByteBufByteReader
Problem / Solution

We want to implement finagle decoding logic in terms of finagle types
such as ByteReader. Here we introduce an abstract implementation of
ByteReader that reads values directly from a netty4 ByteBuf and a
concrete  refinement that guarentees all emitted values are not tied
to the lifetime of the ByteBuf.

RB_ID=914924