finagle-netty4: Do not leak Netty 4 byte buffers into the application code
Problem

While `ByteBufAsBuf` provides a reasonably cheap conversion from Netty 4 ByteBuf
to Finagle Buf it may not be a necessarily safe one to use. Because there is no way
to tell if a user is going to store this `ByteBufAsBuf` for a future need and then
serve it back through a Finagle server (that is going to release the buffer once
it's written), it's not safe to assume the buffer will stay accessible as long as
it's reachable. This may break lots of assumptions about JVM memory management

Solution

Make ByteBufAsBuf smart constructor even smarter such that it produces `Buf.ByteArray`s
when possible. This not only prevents Netty 4 buffers from leaking into the application
code (given our current setup with `AnyToHeap`), but also makes sure to use our fastest
`Buf` implementation when possible.

RB_ID=915022