finagle-core: Reclaim idle sessions when using aperture
Problem

It is possible for sessions to fall out of an aperture after
being used and become idle. For example, during a rolling restart
this will certainly happen.

Solution

Detect if a client is using aperture (and doesn't have a session
idle time set) and set one in this case. The idle time is calculated
as a multiple of the aperture's smooth window. The rationale is
that as the number of smooth windows that pass increases, and a
session remains idle, our degree of certainty that it will remain
idle also increases.

RB_ID=914962