finagle-http2: Fix out of order client stream ID bug
Problem

A race condition exists wherein clients may submit requests with out of order
stream IDs, causing protocol errors. Also, child transports can find themselves
in a state where the parent has placed the final message in an h2 stream on the
child's queue, but the child has not read it yet. This means the child is
unaware that the stream is now half closed.

Solution

Instead of having ChildTransports eagerly obtain new stream IDs on creation and
stream completion, defer stream ID assignment until the first write on a new
stream. Additionally, this must happen in tandem with the call to the underlying
transport's write and be synchronized on the parent MultiplexedTransporter.

Also, ChildTransport now has a new method, offer(HttpObject), which checks to
see if the message signals the end of a stream before queuing.

Result

Regardless of when a ChildTransport is created, it receives a stream ID
atomically with the first write of a new stream, ensuring that all stream IDs
are given out in order. Also, ChildTransports are now immediately aware that the
last message of a stream has been delivered.

RB_ID=914487