finagle-http2: Respect http/2 spec when handling messages with bad stream IDs
Problem

When receiving HTTP/2 frames with stream IDs that aren't recognized, finagle
only logs debug messages. According to the spec, an RST may be sent when an
incoming message is for a stream that has been closed and if enough time has
passed to assume that the peer is misbehaving (opposed to receiving in-flight
messages on a stream that was recently closed). If an incoming message is
received for an idle stream, a GOAWAY/PROTOCOL_ERROR should be sent.
Additionally, the spec states that an endpoint must not send a RST in
response to an RST to avoid a loop.

Solution

If no child exists for a stream, send RST(badId, error=STREAM_CLOSED) if the
stream has been closed. Send GOAWAY(lastId, error=PROTOCOL_ERROR) if the
stream is idle (doesn't exist yet)

RB_ID=915507