finagle: Upgrade to Netty 4.1.10-Final
Problem / Solution

Netty 4.1.10-Final has several features and fixes that we'd like to use, so
this upgrades netty. Additionally, Netty typically recommends a particular
tcnative version, so we bump that to 2.0.1 as well. Additionally, the new
version of netty removes DnsNameResolverBuilder#nameServerAddresses, so
that functionality has been re-created.

RB_ID=916056