util-core: ByteReader extends AutoClosable
Problem

The ByteReader represents an intrinsically mutable representation
of raw data with an implicit notion of owner that codec
implementations can consume. However, it doesn't contain a notion of
resource management making it difficult to represent the semantics
of data sources like the Netty 4 ByteBuf which needs to manage a
reference count.

Solution

Make ByteReader extend AutoClosable to give ByteReader a hook for
managing underlying resources.

RB_ID=916086