finagle-core: bind lb Node type at construction
Problem

The Node type is typically defined by a load metric. In which case,
only the `load` and `pending` fields are needed. However, different
balancers may need to attach additional data to the Node type.  For
example, Aperture needs a `token` to sort the nodes. This was added
to the `NodeT` trait but only used by Aperture forcing the load
metrics to know the internal details of Aperture.

Looking forward, we will want to refine the `NodeT` type to implement
expiration but don't necessarily want these fields directly on `NodeT`.

Solution

Load metrics now can refine the `NodeT` type, but the final binding
is left to the implementation. That is, each implementation has to
implement the `Node` type, `newNode` and `failingNode`.

RB_ID=915772