finagle: DefaultTimer is now service-loaded Problem
Historically, Finagle's default timer was Netty 3 `HashedWheelTimer` and was
defined in `finagle-core`. This makes it difficult to decouple `finagle-core`
from Netty 3. We need a default timer implementation w/o Netty dependency.

Solution

Given that all Finagle protocols are running Netty 4 with Netty 4 `HashedWheelTimer`
by default, we can safely service-load the timer and fallback to just `JavaTimer`
when no timers are available on the classpath.

RB_ID=915924