finagle-core: Manage endpoint closes entirely in TrafficDistributor
Problem

With the introduction of TrafficDistributor, the life cycle of a
load balancer instance changed. In particular, each weight class
receives a new load balancer instance and weight classes are usually
transient. However, the endpoints that back a load balancer instance
have a separate and distinct life cycle. In the past, we managed
this by creating close gates which only open when the endpoints can
be closed as well. However, with the advent of expiring endpoints
(e.g. in aperture), close gates have proven to be rigid and inflexible.

Solution

Lift the management of endpoints into TrafficDistributor entirely.
This is semantically correct since the TrafficDistributor also
creates them.

Result

Endpoints can now be closed without needing to manage gates.

RB_ID=916415