finagle-core: Introduce EndpointFactory
Problem

We would like to unify our root representation of an endpoint
in the load balancer code. We have several instances today
that manipulate or augment an endpoint stack (e.g. via laziness
and address).

Solution

Create a named class that encompasses both of these and is created
by the TrafficDistributor.

RB_ID=916467