finagle-core: Refactor SSL Connection Handlers for Netty 3
Problem / Solution

Work is scheduled to change the SSL / TLS connection handlers for
Finagle running on Netty 3. This is more difficult than necessary
because of where these items are located, how they are named, and how
they are structured. This change renames these items and structures them
in accordance with how the corresponding items are in Finagle running on
Netty 4.

RB_ID=916932