finagle-core: introduce service-loaded NameInterpreter
Problem:

There is no way to change the default `NameInterpreter` except for mutating a
global variable. This is inconvenient if there isn't a clear lifecycle phase to
do it.

Solution:

Provide a `LoadedNameInterpreter` that loads a `NameInterpreter` from the
classpath. It is an error if more than one is available. If none are found,
falls back to `DefaultInterpreter` (this makes it backwards-compatible).

RB_ID=917082