Adding getWithFlag and getsWithFlag apis for Memcache client
Summary: Cache team is implementing hot key detection feature, and we intend to use the flags in the protocol to send hints on what keys can be cached on the client side. Today, get and gets commands drop the flags in the middle of the stack. We want to propagate it to the client, but via 2 new APIs, getWithFlag and getsWithFlag.

As per discussion with Jillian, added tests to MemcachedTest.scala for the 2 APIs.

Basing it off of Kevin's initial change - https://phabricator.twitter.biz/D53083

JIRA Issues: CACHE-8279

Differential Revision: https://phabricator.twitter.biz/D55210