finagle-core: expire endpoints outside of aperture
Problem

Endpoints can fall out the aperture window and become idle.
In these cases, we would like to reclaim these resources.

Solution

Implement expiration on the Node type inside the Aperture load
balancer, which respects the aperture window. That is, nodes that
become idle are only expired iff they are also outside the aperture
window.

We set the idle time as a function of the aperture's smooth window.
Aperture changes are dampened by this window so after X windows
have passed, we can be sufficiently confident that an idle session
is no longer needed. We choose a default of 10 for X which should
give us a high degree of confidence and, based on the default smooth
windows, should be on the order of minutes.

Result

The number of active endpoints (and sessions in multiplexed protocols)
continually converges towards the aperture size.

RB_ID=916508