finagle-memcached: Don't create unnecessary Seqs when encoding
Problem

When encoding a command with associated data, we create
a Seq of 4 elements: name, key, flags, and expiry, and
iterate over this seq to encode each element. It's
unnecessary to create this Seq.

Solution

Save an allocation by passing the name, key, flags, and expiry
to the encoding function directly.

RB_ID=917552
TBR=true