finagle-core: ensure Aperture's expired counter is accurate
Problem

We expire an endpoint and then we don't reset the `idleTime`.
In which case, on subsequent ticks of the expiry clock with no
request activity, we will increment the `expired` counter while
`remake` is a nop.

Solution

Reset the `idleTime` to the sentinel value when we expire.

RB_ID=917715