finagle-netty4: Support passing a chain file for mutual authentication
Summary: Problem

It's impossible to pass a chain file to the client, so the server needs to pass
the intermediate certificates if it wants the client to recognize a root that it
trusts through an intermediate certificate.

Solution

Add support for the CertKeyAndChain API in netty 4, which makes it possible to
pass a chain file to the client.

JIRA Issues: CSL-4686

Differential Revision: https://phabricator.twitter.biz/D59531