finagle-core: change the default load balancer address ordering to be based on a hash of the octets
Summary: Problem

We use the Aurora shard id to create a consistent endpoint ordering
across process boundaries. Since we roll clusters in shard id order,
the rationale was that when using Aperture with a deterministic
ordering, it would be easier to avoid failure if we remove an entire
contiguous slice of the Ring. However, it turns out that this makes
it more difficult for clients to be resilient to rolls because
clients walk the Ring sequentially when they encounter a down node
and each client has to determine liveness independently.

Solution

Change the default address ordering to be based on a MurmurHash3
of the address octets.  Additionally, remove the ability to set an
address ordering globally as it isn't necessary and the per client
granularity should be sufficient.

Result

Shards are uniformly distributed across an Aperture's ring but still
deterministic across process boundaries.

JIRA Issues: CSL-4724

Differential Revision: https://phabricator.twitter.biz/D60698