Don't rely on extending class to pass the `Promise` around.
This changes the responsibilities a bit to ensure that the `Promise[Rep]` is passed around correctly. It is no longer the extending classs' responsibility to do so.

The downside is that for simple cases, there's nothing else to add to the queue, so we have an unecessary instantiation of a `Pending` type. It seems better to ensure correctness than to avoid this instantiation.

Also adds some documentation.