finagle-netty4-http: Aggregate messages that won't have a body
Summary: Problem

The Netty4 based Finagle HTTP client with streaming enabled will emit a
chunked response even in the cases where the response may not have a
body such as responses with status code 204 (No Content).

Solution

We aggregate this class of responses in order to consume the Netty4
LastHttpContent and emit a non-chunked finagle response.

JIRA Issues: CSL-4749

Differential Revision: https://phabricator.twitter.biz/D63417