finagle-core: Reduce variance in Aperture's LoadBand controller
Summary: Problem

The current configuration for the aperture controller leaves room
for clients to diverge since the load band is relatively large.
With a `lowLoad` of 0.5 and `highLoad` of 2.0, we allow the average
load to exists across a range of 1.5.  For example, given a total
offered load of 15, the aperture could be as high as 30 and low as
8 (with no other min/max constraints). In practice, we don't find
such a broad range since our offered load across instances is
relatively uniform (and the controller makes small, discrete
adjustments), but we do see some distribution. This is especially
problematic when using d-aperture since we want a uniform aperture
across clients.

Solution

Tighten the high and load bands and extend the averaging window.
We want the controller to converge on an average load of 1.0 per
aperture serving unit  so we narrow the range to 0.25 (low 0.875,
high 1.125). There are other control algorithms that would allow
us to express this (e.g. a PID controller), but for the time being
the current load band algo is simple and the tuning has proven to
work well with the tests I've run. Additionally, we add some
instrumentation to the controller so we can better observe how it
behaves in production.

Result

In tests, we see distributed clients converge on a uniform aperture
where they once would diverge. Additionally, we tested this with a
prod service and, as per the new instrumentation, the control
algorithm works as expected.

JIRA Issues: CSL-4784

Differential Revision: https://phabricator.twitter.biz/D66795