twitter-server: /admin/ redirects to /admin
Summary: Problem

Since /admin is a portal that points to several other /admin/ pages, it seems
like it should act like a directory, where it should be able to be ended with a
/.  However, today /admin/ doesn't point anywhere.  Some of our New York users
are especially distressed by this.

Also, it turns out that the TwitterServer HttpMuxer wasn't behaving as expected,
because it combined two HttpMuxers that didn't have knowledge of each other.
The TwitterServer HttpMuxer should behave correctly now.

Solution

Redirect /admin/ to prevent some suffering.

JIRA Issues: CSL-4771

Differential Revision: https://phabricator.twitter.biz/D66569