finagle-netty4: Add visibility into slow or blocked Timer tasks
Summary: Problem

The Netty timer should not have blocking tasks on it, much like the
usual for the Netty event loop. However, service owners lack visibility
into what is blocking or "sluggish" for their Timer when it occurs
without happening to catch a stacktrace at that moment.

Solution

Introduce some visibility with a new "finagle/timer/slow" metric
that counts the number of times it has observed a timer task running
for more than 2 seconds. All of the stack traces are also logged,
but to limit output, this will be limited to at most once in a reasonable
time period. Note that the slow checks are done inline with scheduling new
work so as to avoid creating yet another thread. This does imply that
new work must be created in order to trigger this code path.

Result

Service owners get more visibility.

JIRA Issues: CSL-4825

Differential Revision: https://phabricator.twitter.biz/D69274