finagle-mysql: Added the ability to set Isolation Level on transactions
Summary: Problem

There are 4 transaction isolation levels supported by MySQL innodb:

- READ UNCOMMITTED
- READ COMMITTED
- REPEATABLE READ
- SERIALIZABLE

The global default at Twitter is READ COMMITTED, but it is sometimes necessary
for applications to override this based on application logic. Specifically, READ
COMMITTED may allow rows committed by another transaction after your transaction
has started to be seen by your transaction, and this can cause problems if your
logic depends on a higher level of isolation.

Solution

Added an IsolationLevel trait with 4 case objects, and added a new
Transactions.transactionWithIsolation function so the level may be specified
when executing a transactional lambda.

Result

The specified isolation level will only apply to that single transaction, and
not to the session or global environment.

Differential Revision: https://phabricator.twitter.biz/D68944