Finagle ServerSet2 tests: Filter out extraneous identities more easily accept new Identities
Summary: Problem: IdentityTest is very strict about exactly which identities it expects and their order.
This is somewhat brittle in that it prevents adding new identities, especially in lower layers of the stack.

Solution: Filter out additional identities, leaving just the two the test expects to see, so that it can
validate both that those two identities appear, and that they appear in the order expected, without
preventing addition of other Identities later on.

Differential Revision: https://phabricator.twitter.biz/D70690