finagle-memcached: Handler for decoding using a decoder that takes a ByteReader
Summary: Problem

We want to make finagle-memcached's FramingDecoder take a ByteReader.
In order to do this, we need a handler that calls a decoder with a
ByteReader wrapping the Netty ByteBuf in the pipeline.

Solution

Create a ByteReaderDecoderHandler that wraps the Netty ByteBuf in a
ByteReader and calls the decoder. Because the current FramingDecoder
does not satisfy this API, also added a ByteReaderDecoder class
temporarily to make it compile. This will be removed once FramingDecoder
is changed in a future commit.

TBR=true

Differential Revision: https://phabricator.twitter.biz/D70341