finagle-netty4: Fix for BufAsByteBuf.process throwing an exception from Netty
Summary: Problem

`BufAsByteBuf.process` can throw an IndexOutOfBoundsException when
we pass a non-zero start index. ByteBuf.forEach, which `process` calls,
takes an offset and length, and the length we pass can cause the slice
to exceed the length of buf. We define `length` as the minimum of
`readableBytes` and (`until` - `from`). So, a valid value of `length` is
`readableBytes`, but when ByteBuf.forEach iterates over the slice,
it goes from `offset` to (`offset` + `length` - 1). When `offset` is non-zero,
(`offset` + `length` - 1) exceeds the last index of the buffer and throws the
exception.

Solution

Make length the minimum of (`length` - `from`, `until` - `from`). `length` - `from`
is the number of bytes left in the buffer when iterating from `offset`.

JIRA Issues: CSL-4867

Differential Revision: https://phabricator.twitter.biz/D71246