finagle-http2: Stop leaking queues when we close ChildTransport
Summary: Problem

We clean up behind ourselves properly when we finish a request cleanly, but when
we mark a stream as dead, it never gets removed from the hashmap that keeps
track of them.

Solution

Make sure we always remove a stream from the hashmap when it's marked as dead.

JIRA Issues: CSL-4874

Differential Revision: https://phabricator.twitter.biz/D71766