finagle-http: Scope dispatcher stats under client label
Summary: Problem

Dispatcher stats are not scoped under the client label.  This is inconsistent
with all other client stats.

Solution

Modify the endpointer module to use the StatReceiver from the Stack.Params
passed into the module rather than the Stack.Params set on the StackClient.
These are mostly identicial except that `newClient` scopes the StatsReceiver
with the label and passes it into the Stack but does not update the
Stack.Params on the StackClient.

Result

Dispatcher stats are scoped under the client label.

Signed-off-by: Jordan Parker <jparker@twitter.com>

Differential Revision: https://phabricator.twitter.biz/D72265