util-core, finagle-netty4: Add `process` function to ByteReader
Summary: Problem

When decoding bytes in the Netty pipeline using a ByteReader, it
is necessary for some protocols (e.g., Memcached) to be able
to "peek" at bytes in the underlying buffer without advancing
the reader byte cursor. This can be accomplished by adding
a `process` function to ByteReader.

Solution

Add a `process` function to ByteReader. This will be used in
finagle-memcached (in future commits, when the decoder decodes
directly from a ByteReader) to determine the number of bytes before
a line end, and extract bytes from the ByteReader appropriately.

JIRA Issues: CSL-4844

Differential Revision: https://phabricator.twitter.biz/D69785