finagle-memcached: bytesBeforeLineEnd function that takes a ByteReader
Summary: Problem

When the FramingDecoding in finagle-memcached is changed to take a ByteReader
instead of a Buf, we need a way to determine the number of bytes before a newline
in order to slice appropriately.

Solution

Add a `bytesBeforeLineEnd` function to Framer that takes a ByteReader. We already have
a version of that function that takes a Buf, so it's been copied. We can remove the Buf
version once FrameDecoder no longer needs it. Tests for the ByteReader version have
been added to FramerTest (there aren't any tests for the Buf version, but there's no point
adding them now).

JIRA Issues: CSL-4852

Differential Revision: https://phabricator.twitter.biz/D70440