finagle-core: Pass in Stack params in LoadBalancerFactory
Summary: Problem

It useful for load balancer implementations to have access to the entire
stack context. For example, we currently pass in a timer to the aperture
balancers instead of using the stack's timer because we don't have access
to it.

Solution

Change the API for LoadBalancerFactory#newBalancer to take Stack.Params.

TBR=true

Differential Revision: https://phabricator.twitter.biz/D73129