finagle-core: Add retry on throws ResponseClassifier
Summary: Problem

During the migration to MethodBuilder API for some clients, we
discovered that we are using the same code in multiple files.  We felt
it would be good to port the reusable code into its own separate
constant (DRY).

Solution

Added the classifier constant, added the unit test for the constant.

Result

Ease of development, improved reusability of code.

JIRA Issues: CSL-4849

Differential Revision: https://phabricator.twitter.biz/D70475