finagle-core: Export tls information to the R* Registry
Summary: Problem

We don't have a good way of telling how a client or server is configured
tls-wise.

Solution

Make sure that clients and servers export their tls parameters to the R*
Registry.  For the server, there wasn't a great place to do it so I introduced a
generic `Dispatcher` module which doesn't do anything.

JIRA Issues: CSL-3028

Differential Revision: https://phabricator.twitter.biz/D72766