finagle-core: log aperture contents
Summary: Problem / Solution

It's useful to inspect the contents of a clients aperture for verification and debugging.
There isn't an easy way to export this data via stats so we opt to log the contents
of the aperture. We log on each distributor rebuild – rebuilds are not frequent and
usually concentrated around events where this information would be valuable (i.e.
coordinate changes or host add/removes). To be sure that we don't flood the log,
we only log a slice of size min(aperture, 100).

JIRA Issues: CSL-4796

Differential Revision: https://phabricator.twitter.biz/D74064