finagle-mux: Reduce synchronization points in TagMap/TagSet
Summary: Problem / Solution

Both TagMap and TagSet do extensive synchronization on `self` to provide
memory safety, but much of it isn't necessary. To remove some of it, we
make TagSet an implementation detail of TagMap and remove TagSets
internal synchronization, relying on TagMap to manage access safely, which
it already does.

JIRA Issues: CSL-4918

Differential Revision: https://phabricator.twitter.biz/D74207