finagle-netty4: Extract general parts of Netty4Listener
Summary: Problem / Solution

Netty4Listener is specialized to the Transport model, but the majority
of the logic is not specific to Transport. To enable the construction of
Listeners that don't use the Transport model we should extract the
generic parts which are everything but the installation of the
ServerBridge.

JIRA Issues: CSL-4921

Differential Revision: https://phabricator.twitter.biz/D74426