finagle-netty4: Extract generic parts of Netty4Transporter
Summary: NOTE: Unfortunately it's difficult to see, but the vast majority of this change is moving the `Netty4Transporter.build` logic into the helper class, copy paste. The difference is in how the Promise is satisfied, and it has been made into a generic type `T` instead of a `ChannelTransport[In, Out]`, which is not refined at the use site in `Netty4Transporter`. The builder returns a `Future[T]` instead of a raw `T` to facilitate the push model. This allows extracting a `Service` from the `ServiceFactory` (a Future) and initializing the session before the pipeline starts to get data (it waits until the `Future` has resolved).

Problem / Solution

Much of the logic in Netty4Transporter is not specific to the
Transporter model so to prepare for the experimental push-based API we
extract the majority of it so that it can be shared.

JIRA Issues: CSL-4935

Differential Revision: https://phabricator.twitter.biz/D75423