finagle-memcached sbt test changes
Summary: Problem

finagle-memcached's tests depend on finagle-core's `StringClient`
and `StringServer`. These classes require a dependency on
finagle-netty3 code, which finagle-memcached was missing.

Solution

Add a test dependency on finagle-netty3 to finagle-memcached before
the dependency on finagle-core, so that the correct set of classes is
included when running tests with SBT.

JIRA Issues: CSL-5011

Differential Revision: https://phabricator.twitter.biz/D78865