finagle-stats: Only assign verbosity at metric's creation
Summary: # Problem

There is a bug when verbosity level of a gauge is only assigned when it's being queried/created for
the second time, which never happens in practice (thanks to cumulative gauges in SR).

# Solution

Always assign the verbosity of any given metric once, when it's created.

JIRA Issues: CSL-4916

Differential Revision: https://phabricator.twitter.biz/D79337