finagle-core: Fix Failure helper functions
Summary: Problem / Solution

Failure.wrap will add the `Wrapped` flag to a Failure even if we are only
adding the flag to a Failure which was passed in, which may not be wrapped.
The Failure.retryable function doesn't ensure that the passed Throwable
isn't already flagged `NonRetryable`.

JIRA Issues: CSL-5040

Differential Revision: https://phabricator.twitter.biz/D79993