finagle-stats: Exporting correctly the values stored in long gauges.
Summary: Problem:

An unintended side-effect of https://phabricator.twitter.biz/D73497
was that the values of long gauges are now exported as floats. So when
a gauge is set to a large value, the exported value loses precision,
which incorrectly triggers alerts.

Solution:

Add a registerLongGauge() method that mirrors the functionality of
registerGauge(), but exports values as longs instead of floats.

JIRA Issues: CSL-5068

Differential Revision: https://phabricator.twitter.biz/D82059