finagle-http2: MultiplexedTransporter uses single executor
Summary: Problem

MultiplexedTransporter is complicated and currently requires extensive
synchronization. Additionally, satisfying promises in synchronized blocks is
undesirable because it can lead to deadlocks. This leads to added complexity and
the possibility of race conditions.

Solution

Switch to a single executor model where the critical sections are guaranteed to
run one at a time in calling order without synchronized blocks.

Differential Revision: https://phabricator.twitter.biz/D82072