finagle-core: Fix ServerBuilder test flakiness
Summary: Problem

The registry entries for a server contains the entry
Entry(List(server, string, test, Listener), Netty4Listener).
This entry does not contain the bound address, but does
start with the "server, string, test" prefix, so the test that
looks for the bound address fails when this entry is the head
of the server entries list and we look for the bound address
in the head element.

Solution

Also filter by entries that contain an address when we construct
the list of server entries.

JIRA Issues: CSL-5120

Differential Revision: https://phabricator.twitter.biz/D84316