finagle-core: support fractional apertures when using d-aperture
Summary: Problem

When we restrict apertures to positive integers, we can not evenly cover the
ring without a sufficiently large aperture. In particular, the aperture must be
the least common multiple between the number of servers and clients. Indeed, this
does not scale, and in some cases (i.e. # of servers and clients are relatively prime),
we need the aperture to be fully open.

Solution

Add support for fractional apertures. That is, clients can talk a portion of a server.
This allows us to naturally express full ring coverage – each client must simply talk
to the servers which fall within its portion of the ring. However, it does require
that our aperture increase/decrease by the ring unit of a client. For example, if
a client covers 3.33 of the server ring, then we need to increment the aperture by 3.33.

In terms of implementation, `Ring` now  implements a `pick2` method which picks
two indices, without replacement, over a  specified range. We remove the notion
of ring order, since it is implicitly and naturally conveyed as part of the pick.
This does, however, mean we need to expose the indices that are part of the aperture
so that features like expiration can work correctly.

JIRA Issues: CSL-5063

Differential Revision: https://phabricator.twitter.biz/D84452