finagle-mux: MuxFramer handles discards correctly
Summary: Problem

On receiving a discard message, the MuxFramer is checking its tagMap for
the wrong tag so it could end up prepending existing fragments of the
associated dispatch message. Additionally, if it had been using the
right tag, it would have swallowed the message and the associated
Rdiscarded would have been lost to the client due to it being not
propagated in the client case or not sent in the server case.

Solution

Use the right tag and synthesize/propagate the Rdiscarded message as
appropriate.

JIRA Issues: CSL-5149

Differential Revision: https://phabricator.twitter.biz/D87096