finagle-mux: Fix tag collision in ClientDispatcher
Summary: Problem

Interrupting a dispatch can result in a tag collision where the
interrupt can accidentally fail a subsequent dispatch, not the one that
the interrupt belongs to.

Solution

Change the synchronization scheme of the TagMap such that its the
ClientDispatchers responsibility, making it simple to perform a 'compare
and set' like operation in one atomic action. This lets us check if the
Promise in the TagMap is the one associated with the dispatch, and if
not, we leave it alone.

JIRA Issues: CSL-5167

Differential Revision: https://phabricator.twitter.biz/D87933