finagle-core: refactor logical aperture calculation
Summary: Problem

While calculating the aperture in terms of our unit space [0, 1], we were
jumping from the unit space to the integral space. This was more complex
than needed and caused certain issues around the edge case where we need
a width of 1.0.

Solution

Calculate the aperture while remaining in the unit space and bound the
`Coord.width` to be between [0, 1] to avoid any floating point precision
errors.

Differential Revision: https://phabricator.twitter.biz/D88180