finagle-mux: Ensure we actually turn on tls
Summary: Problem

We don't have a good way of checking that we actually enable TLS.

Solution

Pass a function as a param to change how we enable TLS, so that we can record
the bytes that are sent over the wire and be sure that we're turning on TLS
successfully.  This has the side benefit of regularizing how to tls-ify the
handler between the client and server.

JIRA Issues: CSL-5143

Differential Revision: https://phabricator.twitter.biz/D87948