finagle-core: ignore user-defined minAperture when using d-aperture
Summary: Problem

Setting a `minAperture` serves two primary purposes:

1. Ensure that the respective client is resilient and has enough
    warm sessions in case one goes down. Essentially, we want to
    ensure the client has more than one in this case.

2.  Mitigate the effects of the load distribution when using random aperture.
     That is, by expanding the min, it allows each client in a distributed system to
     talk to more of the backends and tighten the load distribution.

Both of these are obsolete when using d-aperture. #1 is solved since we already
set a min of `4` in order to respect node weights and #2 is the chief problem
which d-aperture is designed to solve.

Solution

Ignore the user-defined `minAperture` when using d-aperture. This will
additionally allow us to toggle this feature on without having to worry
about stale user configuration.

Differential Revision: https://phabricator.twitter.biz/D88848