finagle-http2: Upgrade to netty 4.1.15.Final
Summary: Problem

Finagle's http/2 implementation relies heavily on netty's, which has several
bugs.  Some of those bugs are fixed in 4.1.15, but it changes the API
substantively too.

Solution

Upgrade finagle's usage to the new API.  The new API also introduces some new
bugs, so we have two work-arounds for them.  There's a reference-counting bug at
upgrade time in h2c upgrades, and there's a null pointer exception, also at h2c
upgrade time.  The work-around for the reference-counting is to add in an extra
increment, and for the null pointer exception, to set the fields that are unset.

https://github.com/netty/netty/issues/7172
https://github.com/netty/netty/issues/7173

Result

Finagle's http/2 implementation is less buggy--in particular, it can tell more
when something has been written successfully or not.

JIRA Issues: CSL-5130

Differential Revision: https://phabricator.twitter.biz/D87685