finagle-http2: Revert "finagle-http2: Upgrade to netty 4.1.15.Final"
Summary: Problem

We located this as the root cause of slightly increased StreamClosed errors on
inter-service communication and elevated exceptions during service deploy.

Solution

Perform a revert as a temporary fix (this branch) because of the impact on
deploy. Investigation and an actual fix will be done separately.

JIRA Issues: CSL-5202

Differential Revision: https://phabricator.twitter.biz/D91202