finagle-netty3: Remove numWorker test
Summary: Problem / Solution

The numWorkers was dependent on the host having 4 or more
processors. This caused a failure on a recent Travis CI run
where it had less than 4 processors. Removing the test as it
doesn't add much value nor is it worth copying the business
logic into the unit test.

JIRA Issues: csl-5207

Differential Revision: https://phabricator.twitter.biz/D91420