finagle-mysql: ability to toggle CLIENT_FOUND_ROWS flag
Summary: Problem

Original issue: #637

Unable to change the CLIENT_FOUND_ROWS flag. The flag (which
finagle-mysql enables by default) changes the behavior of
affected-rows for UPDATE and INSERT ... ON DUPLICATE KEY UPDATE
statements.

From the manual (v5.7):

For UPDATE statements, the affected-rows value by default is the
number of rows actually changed. If you specify the CLIENT_FOUND_ROWS
flag to mysql_real_connect() when connecting to mysqld, the
affected-rows value is the number of rows “found”; that is, matched by
the WHERE clause.

For INSERT ... ON DUPLICATE KEY UPDATE statements, the affected-rows
value per row is 1 if the row is inserted as a new row, 2 if an
existing row is updated, and 0 if an existing row is set to its
current values. If you specify the CLIENT_FOUND_ROWS flag, the
affected-rows value is 1 (not 0) if an existing row is set to its
current values.

Solution

Added withAffectedRows() on Mysql.client which when used removes the
FoundRows capability when connecting to the server.

My implementation adds another argument to the Handshake case class to
follow how ConnectWithDB is handled. I've also considered checking for
the param in Handshake.apply and if added passing it with
Capability.baseCap - Capability.FoundRows.

Signed-off-by: Kevin Oliver <koliver@twitter.com>

JIRA Issues: CSL-5211

Differential Revision: https://phabricator.twitter.biz/D91406